requirements.txt
future==0.15.2
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.1.2
python-twitter==3.1
requests==2.10.0
requests-oauthlib==0.6.1
